add routes.py and move all the routes in flask to this file. Test implementation
